Make properties grid-aware
- Mouse position from fe.get_input_pos()
- FeText char_size and margin
- FeText and FeListBox outline, bg_outline, glyph_size, msg_width, msg_height and line_height
- FeListBox sel_outline
- FeRectangle outline

+++ src/fe_vm.hpp
@@ -232,7 +232,7 @@ public:
 	static void cb_remove_signal_handler( Sqrat::Object, const char *);
 	static void cb_remove_signal_handler( const char * );
 	static bool cb_get_input_state( const char *input );
-	static int cb_get_input_pos( const char *input );
+	static float cb_get_input_pos( const char *input );
 	static bool do_nut(const char *);
 	static bool load_module( const char *module_file );
 	static void print_to_console( const char *str );
